authentication app features

+++ todo/views.py
@@ -1,9 +1,11 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect
 from .forms import TodoForm
 from .models import Todo
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
 from django.urls import reverse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 
 
@@ -16,9 +18,9 @@ def get_showing_todos(request, todos):
     return todos
 
 
-
+@login_required
 def index(request):
-    todos= Todo.objects.all()
+    todos= Todo.objects.filter(owner= request.user)
     completed = Todo.objects.filter(is_completed=True).count
     not_completed = Todo.objects.filter(is_completed=False).count
     context = { 'todos': get_showing_todos(request, todos),
@@ -29,7 +31,7 @@ def index(request):
     return render(request, 'todo/index.html', context)
 
 
-
+@login_required
 def create_todo(request):
     form = TodoForm()
     context = {'form': form}
@@ -41,34 +43,43 @@ def create_todo(request):
         todo.title        = title
         todo.description  = description
         todo.is_completed = True if is_completed=="on" else False
+        todo.owner = request.user
         
         todo.save()
-        print(todo.id)
+        messages.add_message(request, messages.SUCCESS, 'Task was created successfully')
+        
         return HttpResponseRedirect(reverse("todo", kwargs={'id': todo.pk}))  # reverse because we don't know the id
     return render(request, 'todo/create-todo.html', context)
 
 
-
+@login_required
 def todo_details(request,id):
     todo = get_object_or_404(Todo, pk=id)
+    if todo.owner != request.user:
+        return redirect(reverse('home'))
     context = {'todo': todo}
     return render(request, 'todo/todo-details.html', context)
 
 
-
+@login_required
 def delete_todo(request,id):
     todo = get_object_or_404(Todo, pk=id)
+    if todo.owner != request.user:
+        return redirect(reverse('home'))
     context = {'todo': todo,}
     
     if request.method == 'POST':
         todo.delete()
+        messages.add_message(request, messages.SUCCESS, 'Task was deleted successfully')
         return HttpResponseRedirect(reverse('home'))
     return render(request, 'todo/delete-todo.html', context)
 
 
-
+@login_required
 def edit_todo(request,id):
     todo = get_object_or_404(Todo, pk=id)
+    if todo.owner != request.user:
+        return redirect(reverse('home'))
     form = TodoForm(instance = todo)
     context = {'todo': todo, 'form': form}
 
@@ -82,5 +93,7 @@ def edit_todo(request,id):
         todo.is_completed = True if is_completed == "on" else False
         
         todo.save()
+        messages.add_message(request, messages.SUCCESS, 'Task was edited successfully')
+        
         return HttpResponseRedirect(reverse("todo", kwargs={'id': todo.pk}))
     return render(request, 'todo/edit-todo.html', context)
